fix(filters): RED and OPS lossless parsing and filtering
