+++ internal/domain/filter.go
@@ -475,7 +475,10 @@ func (f *Filter) CheckFilter(r *Release) ([]string, bool) {
 		f.addRejectionF("formats not matching. got: %v want: %v", r.Audio, f.Formats)
 	}
 
-	if len(f.Quality) > 0 && !sliceContainsSlice(r.Audio, f.Quality) {
+	//if len(f.Quality) > 0 && !sliceContainsSlice(r.Audio, f.Quality) {
+	//	f.addRejectionF("quality not matching. got: %v want: %v", r.Audio, f.Quality)
+	//}
+	if len(f.Quality) > 0 && !containsMatchBasic(r.Audio, f.Quality) {
 		f.addRejectionF("quality not matching. got: %v want: %v", r.Audio, f.Quality)
 	}
